+++ Binance.Api/Spot/Socket/BinanceSpotStreamUpdate.cs
@@ -3,11 +3,4 @@
 /// <summary>
 /// User data stream event
 /// </summary>
-public record BinanceSpotStreamUpdate : BinanceSocketStreamEvent
-{
-    /// <summary>
-    /// The listen key the update was for
-    /// </summary>
-    [JsonIgnore]
-    public string ListenKey { get; set; } = string.Empty;
-}
+public record BinanceSpotStreamUpdate : BinanceSpotUserDataStreamEvent;